+++ src/cipher/chacha20poly1305/chacha20poly1305_amd64.go
