+++ src/cipher/poly1305/sum_amd64.s
