+++ src/cipher/poly1305/sum_arm.go
