+++ src/cipher/poly1305/sum_ref.go
